Sprint 11 Исправил замечания по спринту
